==> code/4_melt_tables_PowerBI_integration.ipynb <==
import os
import re
import json
import pickle
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from pprint import pp
from collections import Counter
import seaborn as sns

dirs={}
dirs['project']=os.path.abspath('..')+'\\'
dirs['code']=dirs['project']+'code\\'
dirs['data']=dirs['project']+'data\\'
dirs['models']=dirs['project']+'models\\'
dirs['results']=dirs['project']+'results\\'
dirs['rawresults']=dirs['results']+'raw\\'
dirs['meltresults']=dirs['results']+'melt\\'

First, satisfy the current PowerBI design by exporting one large Destination Lookup Table

df_thresholds = pd.read_pickle(dirs['rawresults']+'similarity_thresholds.pkl')
df_similarity = pd.read_pickle(dirs['rawresults']+'similarity_matrix.pkl')
df_destination = pd.read_pickle(dirs['rawresults']+'destination_lookup_table.pkl')

df_merge = df_destination.drop(columns=['Category original']).merge(df_similarity[df_thresholds.columns], how='left', left_index=True, right_index=True, copy=True)
for t in df_thresholds.columns:
    df_merge['Yes '+t] = df_merge[t].apply(lambda s: 'Yes' if s >= df_thresholds[t].iloc[0] else 'No')
df_merge.to_excel(dirs['results']+'\\Destination lookup table.xlsx')

themes = df_thresholds.columns
df_thresholds = df_thresholds.transpose()
df_thresholds.index.name='Theme'
df_details = df_similarity[['Title', 'Text', 'URL']]
df_similarity = df_similarity.reset_index().melt(id_vars='Topic', var_name='Theme', value_name='Similarity', value_vars=themes).sort_values(['Topic', 'Theme']).set_index('Topic')

df_thresholds.to_csv(dirs['meltresults']+'theme_thresholds.csv')
df_similarity.to_csv(dirs['meltresults']+'topic_theme.csv')
df_destination.to_csv(dirs['meltresults']+'topic_destination.csv')
df_details.to_csv(dirs['meltresults']+'topic_details.csv')

df_destination
